+++ .github/workflows/nightly.yml
@@ -19,7 +19,7 @@ jobs:
           python-version: 3.9
 
       - name: Set version
-        run: echo "VERSION=$(sed -n "s/^__version__ = '\(.*\)'/\1/p" torch_geometric/__init__.py)" >> ${GITHUB_ENV}
+        run: echo "VERSION=$(sed -n "s/^__version__ = '\(.*\)'/\1/p" torch_geometric/glob.py)" >> ${GITHUB_ENV}
 
       - name: Set time
         run: echo "TODAY=$(date +'%Y%m%d')" >> ${GITHUB_ENV}
