+++ torch_geometric/graphgym/contrib/loss/__init__.py
@@ -4,5 +4,5 @@ import glob
 modules = glob.glob(join(dirname(__file__), "*.py"))
 __all__ = [
     basename(f)[:-3] for f in modules
-    if isfile(f) and not f.endswith('__init__.py')
+    if isfile(f) and not f.endswith('glob.py')
 ]
