+++ torch_geometric/nn/__init__.py
@@ -7,7 +7,6 @@ from .to_hetero_with_bases_transformer import to_hetero_with_bases
 from .aggr import *  # noqa
 from .conv import *  # noqa
 from .norm import *  # noqa
-from .glob import *  # noqa
 from .pool import *  # noqa
 from .unpool import *  # noqa
 from .dense import *  # noqa
